Configure cdsapirc file
This is to access the datasets through CDS API(). I was having
trouble trying to get it to work, as it would find an error with
the file every time. Finally, I went through the tutorial again and
it worked. The next step is to do some simple exploratory data ana-
lysis and visualisation.