Select specific variables in subset
One of the largest issues we are going to likely run into is the
one that results in us sending over a JSON object that is huge due
to the data in the subset dataset. This can cause serious issues
like unresponsive visualisations later, so we want to try and select
variables that we are interested in, along a specific time slice,
along a specific longitude and latitude. Then, the idea is that
we convert into a JSON object, use an API to send the dataset to
the front-end, and then use three.js to visualise it. The general
overview is simple, however, actual implementation may prove to be
complex.