Fix datetime object not serializable issue
Datetime objects are not serializable (ie able to be converted to
JSON as it doesn't natively support it), for this reason we needed
to convert the objects into a string format. I first checked what
type of format the variable 'time' was, which was 'datetime64', but
upon running the script, the 'datetime.datetime is not serializable'
error also came up. So I created a custom function that converts
them both to a standardised string format (datetime64 when converted
using str() already comes in an ISO format). If it's neither of those,
we raise a TypeError. Upon running the script, it seemed to work.