Use flask to create API endpoint
We first of all switched to a more modular approach, importing our
view_data module into app.py, creating a dependency graph of it de-
pending on the former. We added the url path that the flask applic-
ation will listen to, and called the getWeatherData function from
the imported module, so that we can view it. At this point we should
be able to fetch the data from the front-end quite easily and get
started with visualising in 3D. We also refactored some code, for
example all the operations for the data_dictionary are all done at
once, which removed a lot of unneeded lines of code. We also can
get rid of the method that handles datetime objects by converting
them to a string, as jsonify from flask handles this internally
anyway.