Add CSS for canvas element to fill entire screen
Some default browser styles were overriden, for example we reduced
the margins to 0, and we hid overflow to prevent any scrolling. We
then targeted a canvas element (our three canvas will get targeted
by this selector) and displayed it as a block that fills the entire
width and height of the screen.