Add earth texture to globe
I don't necessarily like this texture of the earth but I wanted to
see how it is to try and load one. For some reason the canvas went
back to half screen on the left, for which a fix needs to be found
very soon. Also half the globe is black, so that is not exactly
useful. The globe image of the earth was taken from the NASA website.