Convert TIFF file to PNG for high resolution texture
We converted a tiff file format to PNG, most of the quality is ret-
ained, that is because PNG is lossless, so no quality loss in the
conversion itself. Then we just use that texture instead. Although!
I don't know why there is a black side to the earth..? is that bec-
ause it represents day / night? Perhaps. Either way, we need the wh-
ole earth, not just one part of it, so we need to find a fix for that.