Issue with image
I think the issue is that the global spheres represent only an ang-
le of the earth and not all of it. I tried using a rectangular map
and allow three.js to try and map it onto the sphere, but wow, that
was worse than the initial, although no dark part any longer. I may
have to rethink my revolving earth globe idea.