Change sphereGeometry to IcosahedronGeometry
This is done because icosahedronGeometry is made up of equilateral
triangles that are evenly distributed across the surface, whereas
sphereGeometry uses rectangles, so it is a bit less uniform. We've
actually seen before, in one of the commits that sometimes when we
apply a texture to sphereGeomtry it can get stretched in certain
areas. We care about distortion, so IcosahedronGeometry is the be-
tter choice.