Simulate earth axial tilt
We simulated the axial tilt by creating an earth group, which all-
ows us to group multiple objects together, so in the case of our
sphere we are able to group the texture and the shape together and
then change the rotation in the z direction for the entire group.
Thus we simulate the earths tilt in radians.