Remove slicing by latitude in getWeatherData function
The dataset has no latitude values, and so it is meaningless to sl-
ice by the latitude. Instead now we only slice by the longitude. H-
owever, when we run the flask app, the JSON response is massive.
This is because it contains all the metadata and coordinate arrays,
so we might want to return a dataset that is a bit more manageable.
Maybe extracting only the values?