Fix fetching weather data from front-end
Some time back when I was adding what port my app.py flask app will
work on, I had set it to 5000, which was later interpreted to be b-
usy, and so I had to change it to 5001. However, I never changed
the port from which I am fetching in the fetch-request.js module,
and so that explains why it couldn't fetch the data. Now it can.