Fix empty getWeatherData() function dictionary
When running a flask server, the json object returned would be emp-
ty. It was the reason why I thought maybe there is no latitude, be-
cause upon its removal I didn't get epmpty arrays returned anymore.
However, the cause of this issue was not the absence of a latitude
coordinates, but rather that slice expects values between a start
and end. So when I did 30, 50 it did not work because the latitude
coordinates are not in ascending order but rather descending. Swi-
tching the values around therefore fixed the issue.