Add EventListener for window resize
By default, three.js doesn't automatically change the renderer and
the camera aspect when the window is resized. So for this we add an
event listener that fires any time the window is resized, setting the
rendered to the new window height and widithm and updating the camera
projection matrix to make sure our objects are mapped correctly.