Fix issue of plotting points on the earth
The issue has been solved. The problem was simply that for every
latitude we were assuming a correspondent longitude value, which
is incorrect. Rather, for every latitude value we need the entire
length of the longitude values. This will allow us to map the whole
earth. Now, that has been successful but we are running into probl-
ems. Mainly, performance problems. Maybe we could try sending the
data forward to the front-end in streams? I am not entirely sure.
But performance has to be improved or it's going to be too hard