Convert tcc to image texture and apply to globe
We created a flask API endpoint to generate the texture dynamically.
We will try and use 3D noise textures, they are neither too comput-
ationally expensive and neither "truly" 3D. They fake 3D depth,
but if we want to create a web visualisation product that scienti-
sts can use, we want to make it accessible to every hardware. It
should not be that only the computers with good GPU's can run this
tool. We loaded the generated texture and applied it over the earth.
Now we will need to apply a shader to make it look 3D by moving the
clouds, changing noise patterns that evolve in response to user
interaction etc. This is the computationally expensive part, however
we will have natural and realistic visuals to go along with it.