Stack 2D cloud layers to give visual effect of 3D
The issue here is that.. i'm not sure what i was really expecting
from the visualisation. It's definitely based on the ERA5 dataset,
but is this really the best we can do? The dataset itself doesnt
contain variables such as cloud thickness, so Im not sure how I
would visualise it in a 3d volumetric way. It rather has 2D ..
values.