Map low density clouds with light blue and high density with white
Before it mapped based on density as well, but it was white and
the opacity changed. So low density clouds would fade away and the
high density clouds would be flat white. That's not that useful
to a scientist, so now we made it so that low density clouds (ciwc
data) are mapped to a light bluish, and high density clouds are ma-
pped to a bright white. So they are easier to tell apart visually.
Next step is to implement some sort of slicing by pressure level.